Add MFA authncontext response validator
The validator should check if the IdP has the for the IdP SP
combination configured authn method in the response to the SP when
configured.

The authn method could be found in the authncontextclassref or as value
in the multivalued `http://schemas.microsoft.com/claims/authnmethodsreferences`
attribute.

When not found but configured this should result in an exception.

If